Add _attribution module
